requirements.txt
# Core requirements for Cold Drink Detector
ultralytics>=8.0.0
matplotlib>=3.7.0
opencv-python>=4.7.0
pyyaml>=6.0
numpy>=1.24.0
pillow>=9.0.0
